requirements.txt
dash
plotly
python-dotenv
databricks-sql-connector
dash-bootstrap-components